+++ android/app/src/main/kotlin/com/skoryk/base/pacer/WorkoutServiceEvents.kt
@@ -1,4 +1,4 @@
-package com.skoryk.basepacer
+package com.skoryk.base.pacer
 
 object WorkoutServiceEvents {
     @Volatile